LayoutTests/platform/win/fast/frames/frame-with-noresize-can-be-resized-after-removal-of-noresize-expected.txt



--------
Frame: 'testFrame'
--------


--------
Frame: 'results'
--------
This tests that a frame with a source code-specified noresize attribute can be resized after programmatically removing its noresize attribute.

FAIL document.getElementById("testFrame").width should be 210. Was 200.
